refactor: 파일명 통일 및 utils 역할별 분리
- 케밥케이스 → 카멜케이스: supabase-server, menu-policy, day-bar, hero-status
- 케이싱 충돌 해소: dayBar.ts → DayBar.utils.ts, heroStatus.ts → HeroStatus.utils.ts
- utils/utils.ts 분리: cn.ts (클래스 유틸), date.ts (날짜 유틸)
