+++ src/app/api/picks/route.ts
@@ -1,6 +1,6 @@
 import {NextRequest, NextResponse} from 'next/server';
 
-import {supabaseServer} from '@/lib/supabase-server';
+import {supabaseServer} from '@/lib/supabaseServer';
 import {PickResult, PickType} from '@/types/vote';
 
 const PICK_TYPES: PickType[] = ['COURSE_1', 'COURSE_2', 'TAKE_OUT', 'pass'];
