+++ src/components/menu/MenuBoard/MenuBoard.styles.ts
@@ -1,4 +1,4 @@
-import {cn} from '@/utils/utils';
+import {cn} from '@/utils/cn';
 
 export const todayButtonClass = (isActive: boolean) =>
   cn(
