+++ src/components/menu/MenuBoard/_DayBar/DayBar.styles.ts
@@ -1,4 +1,4 @@
-import {cn} from '@/utils/utils';
+import {cn} from '@/utils/cn';
 
 export const navButtonClass = (canGo: boolean) =>
   cn(
