+++ src/store/useDateStore.ts
@@ -2,7 +2,7 @@ import toast from 'react-hot-toast';
 import {create} from 'zustand';
 
 import dayjs from '@/lib/dayjs';
-import {getWeekDays} from '@/utils/utils';
+import {getWeekDays} from '@/utils/date';
 
 interface DateStore {
   today: dayjs.Dayjs;
