+++ src/utils/cn.ts
@@ -1,0 +1,3 @@
+import clsx, {type ClassValue} from 'clsx';
+
+export const cn = (...inputs: ClassValue[]) => clsx(...inputs);
